requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
distlib==0.3.1
Django==3.1.1
django-extra-views==0.13.0
filelock==3.0.12
gunicorn==20.0.4
numpy==1.19.2
pipenv==2020.8.13
psycopg2==2.8.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
ulid-py==1.1.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
